#137972 move extension methods to service
