+++ GovUk.Frontend.Umbraco/Services/IUmbracoPublishedContentAccessor.cs
@@ -1,9 +1,13 @@
-﻿using Umbraco.Cms.Core.Models.PublishedContent;
+﻿using System;
+using Umbraco.Cms.Core.Models.Blocks;
+using Umbraco.Cms.Core.Models.PublishedContent;
 
 namespace GovUk.Frontend.Umbraco.Services
 {
     public interface IUmbracoPublishedContentAccessor
     {
         IPublishedContent PublishedContent { get; }
+        BlockListItem? FindBlockByBoundProperty(string propertyName);
+        BlockListItem? FindBlock(Func<BlockListItem, bool> matcher);
     }
 }
